splitted app code into separate files, added config
