+++ config.py
@@ -1,0 +1,13 @@
+import os
+
+
+basedir = os.path.abspath(os.path.dirname(__file__))
+
+
+class Config(object):
+    SECRET_KEY = os.environ['SECRET_KEY']
+    SQLALCHEMY_DATABASE_URI = os.environ.get('DATABASE_URL') or \
+        'sqlite:///' + os.path.join(basedir, 'app.db')
+    SQLALCHEMY_TRACK_MODIFICATIONS = False
+    # TODO remove after migrating to SQLAlchemy
+    DATABASE = os.path.join(basedir, 'instance', 'advencal.sqlite')
